Added CLI args, better restrictions, fixed robots
Added CLI arguments so multiple crawls can be done in parallel.
Add better url restrictions if we want to restrict excluding subdomain
or whatnot.
Added "http://" to the urls when checking if they're in robots.txt with
the robotparser. The parser won't recognize that the url is included or
excluded otherwise.